+++ context_test.go
@@ -2,6 +2,7 @@ package cli
 
 import (
 	"flag"
+	"sort"
 	"testing"
 	"time"
 )
@@ -19,8 +20,6 @@ func TestNewContext(t *testing.T) {
 	c.Command = command
 	expect(t, c.Int("myflag"), 12)
 	expect(t, c.Float64("myflag64"), float64(17))
-	expect(t, c.GlobalInt("myflag"), 42)
-	expect(t, c.GlobalFloat64("myflag64"), float64(47))
 	expect(t, c.Command.Name, "mycommand")
 }
 
@@ -31,27 +30,11 @@ func TestContext_Int(t *testing.T) {
 	expect(t, c.Int("myflag"), 12)
 }
 
-func TestContext_GlobalInt(t *testing.T) {
-	set := flag.NewFlagSet("test", 0)
-	set.Int("myflag", 12, "doc")
-	c := NewContext(nil, set, nil)
-	expect(t, c.GlobalInt("myflag"), 12)
-	expect(t, c.GlobalInt("nope"), 0)
-}
-
 func TestContext_Float64(t *testing.T) {
 	set := flag.NewFlagSet("test", 0)
 	set.Float64("myflag", float64(17), "doc")
 	c := NewContext(nil, set, nil)
 	expect(t, c.Float64("myflag"), float64(17))
-}
-
-func TestContext_GlobalFloat64(t *testing.T) {
-	set := flag.NewFlagSet("test", 0)
-	set.Float64("myflag", float64(17), "doc")
-	c := NewContext(nil, set, nil)
-	expect(t, c.GlobalFloat64("myflag"), float64(17))
-	expect(t, c.GlobalFloat64("nope"), float64(0))
 }
 
 func TestContext_Duration(t *testing.T) {
@@ -80,30 +63,6 @@ func TestContext_BoolT(t *testing.T) {
 	set.Bool("myflag", true, "doc")
 	c := NewContext(nil, set, nil)
 	expect(t, c.BoolT("myflag"), true)
-}
-
-func TestContext_GlobalBool(t *testing.T) {
-	set := flag.NewFlagSet("test", 0)
-
-	globalSet := flag.NewFlagSet("test-global", 0)
-	globalSet.Bool("myflag", false, "doc")
-	globalCtx := NewContext(nil, globalSet, nil)
-
-	c := NewContext(nil, set, globalCtx)
-	expect(t, c.GlobalBool("myflag"), false)
-	expect(t, c.GlobalBool("nope"), false)
-}
-
-func TestContext_GlobalBoolT(t *testing.T) {
-	set := flag.NewFlagSet("test", 0)
-
-	globalSet := flag.NewFlagSet("test-global", 0)
-	globalSet.Bool("myflag", true, "doc")
-	globalCtx := NewContext(nil, globalSet, nil)
-
-	c := NewContext(nil, set, globalCtx)
-	expect(t, c.GlobalBoolT("myflag"), true)
-	expect(t, c.GlobalBoolT("nope"), false)
 }
 
 func TestContext_Args(t *testing.T) {
@@ -139,25 +98,6 @@ func TestContext_IsSet(t *testing.T) {
 	expect(t, c.IsSet("myflagGlobal"), false)
 }
 
-func TestContext_GlobalIsSet(t *testing.T) {
-	set := flag.NewFlagSet("test", 0)
-	set.Bool("myflag", false, "doc")
-	set.String("otherflag", "hello world", "doc")
-	globalSet := flag.NewFlagSet("test", 0)
-	globalSet.Bool("myflagGlobal", true, "doc")
-	globalSet.Bool("myflagGlobalUnset", true, "doc")
-	globalCtx := NewContext(nil, globalSet, nil)
-	c := NewContext(nil, set, globalCtx)
-	set.Parse([]string{"--myflag", "bat", "baz"})
-	globalSet.Parse([]string{"--myflagGlobal", "bat", "baz"})
-	expect(t, c.GlobalIsSet("myflag"), false)
-	expect(t, c.GlobalIsSet("otherflag"), false)
-	expect(t, c.GlobalIsSet("bogusflag"), false)
-	expect(t, c.GlobalIsSet("myflagGlobal"), true)
-	expect(t, c.GlobalIsSet("myflagGlobalUnset"), false)
-	expect(t, c.GlobalIsSet("bogusGlobal"), false)
-}
-
 func TestContext_NumFlags(t *testing.T) {
 	set := flag.NewFlagSet("test", 0)
 	set.Bool("myflag", false, "doc")
@@ -171,62 +111,6 @@ func TestContext_NumFlags(t *testing.T) {
 	expect(t, c.NumFlags(), 2)
 }
 
-func TestContext_GlobalFlag(t *testing.T) {
-	var globalFlag string
-	var globalFlagSet bool
-	app := NewApp()
-	app.Flags = []Flag{
-		StringFlag{Name: "global, g", Usage: "global"},
-	}
-	app.Action = func(c *Context) error {
-		globalFlag = c.GlobalString("global")
-		globalFlagSet = c.GlobalIsSet("global")
-		return nil
-	}
-	app.Run([]string{"command", "-g", "foo"})
-	expect(t, globalFlag, "foo")
-	expect(t, globalFlagSet, true)
-
-}
-
-func TestContext_GlobalFlagsInSubcommands(t *testing.T) {
-	subcommandRun := false
-	parentFlag := false
-	app := NewApp()
-
-	app.Flags = []Flag{
-		BoolFlag{Name: "debug, d", Usage: "Enable debugging"},
-	}
-
-	app.Commands = []Command{
-		{
-			Name: "foo",
-			Flags: []Flag{
-				BoolFlag{Name: "parent, p", Usage: "Parent flag"},
-			},
-			Subcommands: []Command{
-				{
-					Name: "bar",
-					Action: func(c *Context) error {
-						if c.GlobalBool("debug") {
-							subcommandRun = true
-						}
-						if c.GlobalBool("parent") {
-							parentFlag = true
-						}
-						return nil
-					},
-				},
-			},
-		},
-	}
-
-	app.Run([]string{"command", "-d", "foo", "-p", "bar"})
-
-	expect(t, subcommandRun, true)
-	expect(t, parentFlag, true)
-}
-
 func TestContext_Set(t *testing.T) {
 	set := flag.NewFlagSet("test", 0)
 	set.Int("int", 5, "an int")
@@ -236,21 +120,36 @@ func TestContext_Set(t *testing.T) {
 	expect(t, c.Int("int"), 1)
 }
 
-func TestContext_GlobalSet(t *testing.T) {
-	gSet := flag.NewFlagSet("test", 0)
-	gSet.Int("int", 5, "an int")
+func TestContext_LocalFlagNames(t *testing.T) {
+	set := flag.NewFlagSet("test", 0)
+	set.Bool("one-flag", false, "doc")
+	set.String("two-flag", "hello world", "doc")
+	globalSet := flag.NewFlagSet("test", 0)
+	globalSet.Bool("top-flag", true, "doc")
+	globalCtx := NewContext(nil, globalSet, nil)
+	c := NewContext(nil, set, globalCtx)
+	set.Parse([]string{"--one-flag", "--two-flag=foo"})
+	globalSet.Parse([]string{"--top-flag"})
 
-	set := flag.NewFlagSet("sub", 0)
-	set.Int("int", 3, "an int")
+	actualFlags := c.LocalFlagNames()
+	sort.Strings(actualFlags)
 
-	pc := NewContext(nil, gSet, nil)
-	c := NewContext(nil, set, pc)
+	expect(t, actualFlags, []string{"one-flag", "two-flag"})
+}
 
-	c.Set("int", "1")
-	expect(t, c.Int("int"), 1)
-	expect(t, c.GlobalInt("int"), 5)
+func TestContext_FlagNames(t *testing.T) {
+	set := flag.NewFlagSet("test", 0)
+	set.Bool("one-flag", false, "doc")
+	set.String("two-flag", "hello world", "doc")
+	globalSet := flag.NewFlagSet("test", 0)
+	globalSet.Bool("top-flag", true, "doc")
+	globalCtx := NewContext(nil, globalSet, nil)
+	c := NewContext(nil, set, globalCtx)
+	set.Parse([]string{"--one-flag", "--two-flag=foo"})
+	globalSet.Parse([]string{"--top-flag"})
 
-	c.GlobalSet("int", "1")
-	expect(t, c.Int("int"), 1)
-	expect(t, c.GlobalInt("int"), 1)
+	actualFlags := c.FlagNames()
+	sort.Strings(actualFlags)
+
+	expect(t, actualFlags, []string{"one-flag", "top-flag", "two-flag"})
 }