+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ## 2.0.0 - (unreleased 2.x series)
 ### Added
 - `NewStringSlice` and `NewIntSlice` for creating their related types
+- `Context.Lineage` to get all contexts from current up to global
+- `Context.LocalFlagNames` to get the flag names from *only* the current context
+
+### Changed
+- `Context.FlagNames` now returns all flags in the context lineage
+- `Context.IsSet` now considers the full context lineage
 
 ### Removed
 - the ability to specify `&StringSlice{...string}` or `&IntSlice{...int}`.
@@ -17,6 +23,11 @@
   arguments](https://github.com/codegangsta/cli/issues/355) when the user
   attempted to mix flags and arguments. Given the trade-offs we removed support
   for this reordering.
+- All `Context.Global*` methods, as the non-global versions now traverse up
+  the context lineage automatically.
+
+### Fixed
+- `Context.BoolT` now returns `true` when not found
 
 ## [Unreleased] - (1.x series)
 ### Added